Fix string copying and string comparison in clapack's ilaenv.c
When running roadrunner through valgrind, it found hundreds of errors of the type "Conditional jump or move depends on uninitialised value(s)", all in ilaenv.c.  This turns out to be because for all string copying and comparisons in the file, all local strings are defined to be length 1, despite being explicitly set to other lengths.  This makes it impossible to properly copy and compare the strings in question.  So when the string "DGEQRF" is sent into the function, only the 'D' is copied into a 6-char string, giving "D*****", and then later the *'s are compared to certain key values.  Telling the copy and compare functions that a 6-char string is a 6-char string and not a 1-char string cleans this up.