Update SBMLNetwork and libsbml.
The libsbml update puts us back on the main branch after having lived in an auxilliary branch for quite some time--things are finally fixed there that work well enough that both we and the libsbml team are happy.

It also fixes a bug we found in libsbml in the render implementation.